third_party/WebKit/LayoutTests/http/tests/devtools/startup/sources/debugger/script-formatter-breakpoints-4-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests the script formatting is working fine with breakpoints.


Running: testSetup

Running: testBreakpointSetInOriginalAndRemovedInFormatted
Adding breakpoint.
Formatting.
Removing breakpoint.
Unformatting.
